+++ src/main/java/popcong/app/application/auth/port/in/TokenGenerateUseCase.java
@@ -1,4 +1,4 @@
-package popcong.app.domain.auth.port.in;
+package popcong.app.application.auth.port.in;
 
 import popcong.app.domain.auth.model.AuthInfo;
