+++ src/main/java/popcong/app/application/auth/service/token/AccessTokenGenerateService.java
@@ -1,10 +1,10 @@
-package popcong.app.application.auth.service;
+package popcong.app.application.auth.service.token;
 
 import io.jsonwebtoken.Jwts;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import popcong.app.domain.auth.model.AuthInfo;
-import popcong.app.domain.auth.port.in.TokenGenerateUseCase;
+import popcong.app.application.auth.port.in.TokenGenerateUseCase;
 import popcong.app.infra.config.jwt.JwtProperties;
 
 import javax.crypto.SecretKey;
@@ -12,25 +12,26 @@ import java.util.Date;
 
 @Service
 @RequiredArgsConstructor
-public class RefreshTokenGenerateService implements TokenGenerateUseCase {
+public class AccessTokenGenerateService implements TokenGenerateUseCase {
 
     private final JwtProperties jwtProperties;
     private final SecretKey secretKey;
 
     /**
-     * AuthInfo의 사용자 id로 Refresh Token 생성
+     * AuthInfo의 사용자 id, providerId, email, role, userRole로 Access Token 생성
      * @param authInfo
-     * @return refreshToken 발급
+     * @return accessToken 발급
      */
     @Override
     public String generateToken(AuthInfo authInfo) {
         Date now = new Date();
-        Long expiration = jwtProperties.getExpirationTime().getRefreshToken();
+        Long expiration = jwtProperties.getExpirationTime().getAccessToken();
 
         return Jwts.builder()
                 .setSubject(authInfo.id().toString())
-                .claim("role",  authInfo.role())
-                .claim("tokenType", "REFRESH")
+                .claim("email", authInfo.email())
+                .claim("role", authInfo.role())
+                .claim("tokenType", "ACCESS")
                 .setIssuedAt(now)
                 .setExpiration(new Date(now.getTime() + expiration))
                 .signWith(secretKey)
