+++ src/main/java/popcong/app/application/auth/service/token/RefreshTokenGenerateService.java
@@ -1,10 +1,10 @@
-package popcong.app.application.auth.service;
+package popcong.app.application.auth.service.token;
 
 import io.jsonwebtoken.Jwts;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import popcong.app.domain.auth.model.AuthInfo;
-import popcong.app.domain.auth.port.in.TokenGenerateUseCase;
+import popcong.app.application.auth.port.in.TokenGenerateUseCase;
 import popcong.app.infra.config.jwt.JwtProperties;
 
 import javax.crypto.SecretKey;
@@ -12,26 +12,25 @@ import java.util.Date;
 
 @Service
 @RequiredArgsConstructor
-public class AccessTokenGenerateService implements TokenGenerateUseCase {
+public class RefreshTokenGenerateService implements TokenGenerateUseCase {
 
     private final JwtProperties jwtProperties;
     private final SecretKey secretKey;
 
     /**
-     * AuthInfo의 사용자 id, providerId, email, role, userRole로 Access Token 생성
+     * AuthInfo의 사용자 id로 Refresh Token 생성
      * @param authInfo
-     * @return accessToken 발급
+     * @return refreshToken 발급
      */
     @Override
     public String generateToken(AuthInfo authInfo) {
         Date now = new Date();
-        Long expiration = jwtProperties.getExpirationTime().getAccessToken();
+        Long expiration = jwtProperties.getExpirationTime().getRefreshToken();
 
         return Jwts.builder()
                 .setSubject(authInfo.id().toString())
-                .claim("email", authInfo.email())
-                .claim("role", authInfo.role())
-                .claim("tokenType", "ACCESS")
+                .claim("role",  authInfo.role())
+                .claim("tokenType", "REFRESH")
                 .setIssuedAt(now)
                 .setExpiration(new Date(now.getTime() + expiration))
                 .signWith(secretKey)
